requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
langchain-groq
langchain-tavily
python-multipart
ipykernel
langchain-openai